This is the test class which lays bare the projection bug that I have been facing. I do not have a PR for fixing this.
What I saw is that the label thing is used 2 times in the method and that htat is what this is about.
It seems to have something to do that in logging we want to show the value that is not mapping right.

A concern with that is that other people could face with that implementation is that when the mapped value happens to be a secret value that
should never be in the log files.  That will be in the log files when the secret is not being able to map.
Dat is a security concern with getting values in the exceptions.