This solves the bug in that it will not throw an exception anymore.
However,  this does not give the results I would expect.

Now when mapping a property that itself is not nullable but is part of a projection path should make the
property change to a nullable property.

So if we map MyClass.MyChildObject.MyIntProp of type int to a target property of type int?
Then I would expect the target to be null when MyClass.MyChildObject is null. And now it is 0 becuase the sourcegetter has int
as its returntype.   But given the projection I would expect the returntype of the whole getter expression to be int? instead of int.
And thus the result value to be null == default(int?)  instead of 0 == default(int).